oscbn_parse now pushes an empty bundle onto the stack if the input string is empty.
Most code using this parser expects to find two elements on
the stack after parsing. Before this change, the parser would
just bail on an empty string and leave it there, forcing a check
to see if the size of the stack had changed.

This change means that we're interpreting an empty string
to mean "empty bundle", which, in this context, makes sense
and is consistent with odot (clicking an empty o.compose
box produces an empty bundle).